+++ src/utils/collections/index.ts
@@ -2,4 +2,5 @@
 export * from "./Es5WeakMap";
 export * from "./HashSet";
 export * from "./KeyValueCache";
+export * from "./SortedKeyValueArray";
 export * from "./WeakCache";